Separate XSLT extension functions
